Moved to another file.
Implemented -a flag.
